requirements.txt
torch
transformers
datasets
matplotlib 
seaborn 
pandas
hydra-core